Formalize image tags and provide a build script
Clean up the tag for the images we'll use, also add
a dockerfile for OSC (we'll need that later to add
in keystone and do all the endpoint creation etc).

Clean up the cinderclient image, use debian and
add the entrypoint `cinder`.

Also, as a helper include a bootstrap-images script
rather than building them from Compose.