Simplify a bit and add Keystone work
This adds some tweaks to simplify things a little bit,
and fixes up some db issues we had in our usage. Get
rid of the bootstrap script and just use the Makefile
going forward.

I'm also adding a Keystone Dockerfile and will get that
synced up as an option/alternative to noauth.