Clean up a bunch of stuff, provide a LVM option
This commit cleans up a bunch of unnecessary stuff, and switches our
default Make to use Ocata images instead of source.

In addition we had the external backend stuff working pretty good, now
we've also added in the ability to run a configured LVM backend as well.
Just remember you need to set up the cinder-volumes VG locally on the
host system.