Make it work for iscsi attaches
Problem with default compose is you get an compose bridged network.  This is
great for everything BUT making iscsi attachments on a remote node.  This goofs
things up because the portal you actually communicate with is the host that the
container is running on, but the tgt service reports back the internal IP and
iscsi says "sorry".

Might be able to set tgt iscsi IP, but this work-around does the trick for now.